re-use a global Kithe::Indexer
Traject makes instantiation of an Indexer kind of expensive currently. We get MUCH better performance this way,
and it should be perfectly thread-safe for traject, even used by multiple threads simultaneously.

One downside is it might do weird things with Rails dev-mode class reloading. We'll have to see how it goes and if we need a workaround.