Make kithe_model_type more resilient
As a result of #43, we have a required kithe_model_type column. We were making sure it was set in an initialize callback, but make sure it's set in a before_validation callback too, in case someone accidentally deleted it. We actually did in our app. Hmm, this feature is a bit fragile, but I think it's worth it.