Some very basic support for getting Blacklight to bulk-fetch underlying db models
There's more docs than there is code!

Sorry, no automated rspec for the BL integration at present. For the moment, figuring out how to test a BL integration/plugin in a sane and reliable way is beyond my energy. There should fairly naturally be tests in our actual app (or other using apps) that would fail if this isn't working, since the point is to write app functionality dependent on it.