Allow a derivative to be manually set on a non-stored file, and not be deleted
Our callback logic tries to ensure that "stale" derivatives aren't around, by deleting derivatives if the SHA256 changes.

This meant if you set a derivative on a file before a SHA was calculated, it would be deleted once the asset was promoted and had a SHA assigned.

This was getting in the way of some test scenarios -- it's hard to figure out how to set up tests efficiently for this stuff. So for now, we will allow a derivative set when there is NO sha to persist once a sha is set. The higher level APIs may still raise if you try to create derivatives on a non-promoted file without a digest. That's fine. This change is mainly being done for testing/manual/hacky scenarios.