Don't call set_leaf_representative on save if it's been explicitly set to a new value already
Got in the way of our bulk imports, if we're bulk importing and setting leaf_representative_id, we don't need the logic to look it up to be called. And in some cases, when we were trying to defer enforcing of constraints and set to an id that isn't in db yet -- the automaticc lookup was setting to nil when we were trying to set to a value.