Move Kithe::ConfigBase to ./lib/ so it will not be auto-loaded
This makes it easier to use in working and non-deprecated way in Rails initialization code, which is one of it's main use cases.

Changes to Rails auto-loading/zeitwerk make it so that you should not be referencing auto-loaded code in Rails configuration/initialization phases.

https://github.com/rails/rails/issues/40904

As we now explicitly require kithe/config_base in the kithe/engine file, this change should be transparent and backwards-compatible.