Avoid making a query on every validation
Leads to a sort of n+1 on batch update. We're going to leave the validation in, but only triggered if the parent has actually changed, no need to run the validation otherwise, so we can skip it and avoid the fetch. If it changed, it is probably already loaded anyway? If not, well, we solved some of the problem here.