Fix for Rails 7 breaking recommended single-table-inheritance approach
We were using an approach to Single-Table Inheritance recommended by Rails guide at: https://guides.rubyonrails.org/v7.0/autoloading_and_reloading_constants.html#single-table-inheritance

But this breaks in Rails 7.  There's an Issue in Rails github where contributors are discussing it: https://github.com/rails/rails/issues/45307

It looks like this is considered a bug in the guide-recommended STI approach, rather than a bug in Rails. They haven't yet consensed on an exact solution, but solution here is my best effort taken from what they are arriving at in discussion there.

It does require the STI base class to be explicitly mentioned in an initializer, when before everything was totally automatic. But this engine only has one, Kithe::Model, so this engine can take care of the one this engine is providing, no problem.