Allow attr_json 2.x
we're still allowing 1.x too.  We're not going to update major version of kithe.

This is a bit fast and loose, as one could imagine a downstream app for which updating to attr_json 2 might break it. But kithe isn't super popular, we think we may know all current users, who are fine with it.