Ensure single downloaded file re-used for metadata and before_promotion, with Shrine.plugin :tempfile
If you have lots of `add_metadata` blocks that need a local file, do they all end up downloading file each time? What about before_promotion blocks too?

Looks like in the default implementation, they don't actually multiply download from (eg) S3, becuase the underlying Down::ChunkedIO caches. But they DO make multiple local copies of the Down::ChunckedIO cache. Which could be non-trivial performance implications if you have a LOT of copies and the files is large. And not sure whether this is part of the Shrine/Down contract to rely on this behavior anyway.

The [Shrine tempfile plugin](https://shrinerb.com/docs/plugins/tempfile) is meant to provide tools to handle this. Can we wire it up so the _promotion_ phase will automatically share a tempfile, for all metadata extraction and before_promotion hoooks?

We can. Kind of a weird implementation, but not too bad. And we seem to be testing it succesfully, and it works.

The local app DOES need to do `Shrine.plugin :tempfile` in an initialier for this to work -- if they don't, there are no real problems caused, but you won't get tempfile re-use. Kithe didn't want to do this becuase it is a global Shrine plugin that will effect the whole app, possibly including non-Kithe uses.

Some more explanation and docs are in the PR.